refactor to reuse work for whole execution flow

+++ index.js
@@ -12,12 +12,10 @@ const split_chunk = function (input) {
   }
   return input_list;
 };
-
-// To avoid intermediate big array
 function* convert_hashmap_to_array(obj) {
-  for (let key in obj) yield new Map().set(key, obj[key]);
+  //   for (let key in obj) yield new Map().set(key, obj[key]);
+  for (let key in obj) yield [key, obj[key]];
 }
-
 const shuffle = function (post_mapper) {
   post_mapper.sort();
 
@@ -42,10 +40,29 @@ const partition = function (key_values_array) {
     const batch = key_values_array.slice(i, upper_bound);
     reducer_tasks.push(batch);
   }
-  return { reducer_size, reducer_tasks };
+  reducer_tasks.sort();
+  return reducer_tasks;
+};
+const replacer = function (key, value) {
+  if (value instanceof Map) {
+    return {
+      dataType: "Map",
+      value: Array.from(value.entries()), // or with spread: value: [...value]
+    };
+  } else {
+    return value;
+  }
+};
+const reviver = function (key, value) {
+  if (typeof value === "object" && value !== null) {
+    if (value.dataType === "Map") {
+      return new Map(value.value);
+    }
+  }
+  return value;
 };
 
-const mapreduce_cluster = function (input, map, reduce, callback) {
+const task_mgr = function (input, map, reduce, callback) {
   const tasks = split_chunk(input);
 
   if (cluster.isMaster) {
@@ -54,13 +71,35 @@ const mapreduce_cluster = function (input, map, reduce, callback) {
       var map_count = 0;
       var reduce_count = 0;
       var post_mapper = [];
+      var post_reducer = [];
+      var reducer_task_size;
       const task = tasks[worker.id - 1];
+
       worker.send({ map_task: task });
 
       worker.on("message", (msg) => {
         if (msg.signal === "MapCompleted") {
-          post_mapper = post_mapper.concat(msg.intermediate);
+          post_mapper = post_mapper.concat(msg.mapped);
           map_count++;
+
+          if (map_count === cores) {
+            console.log(post_mapper);
+            //  e.g. { 'a': [ 1 ], 'b': [ 1 ], "c": [ 1, 1, 1 ] }
+            const key_values = shuffle(post_mapper);
+            const reducer_tasks = partition(key_values);
+            reducer_task_size = reducer_tasks.length;
+            for (const task of reducer_tasks) {
+              worker.send({ reduce_task: task });
+            }
+          }
+        }
+        if (msg.signal === "ReduceCompleted") {
+          post_reducer = post_reducer.concat(Object.entries(msg.reduced));
+          reduce_count++;
+
+          if (reduce_count === reducer_task_size) {
+            callback(new Map(post_reducer));
+          }
         }
       });
 
@@ -68,37 +107,52 @@ const mapreduce_cluster = function (input, map, reduce, callback) {
         console.log(error);
       });
 
-      worker.on("exit", () => {
-        if (map_count == cores) {
-          //  e.g. { 'a': [ 1 ], 'b': [ 1 ], "c": [ 1, 1, 1 ] }
-          const key_values = shuffle(post_mapper);
-          const { reducer_size, reducer_tasks } = partition(key_values);
-          let final = {};
-          for (const reducer_task of reducer_tasks) {
-            reducer_task.forEach((pair) => {
-              const twisted = Array.from(pair)[0];
-              final[twisted[0]] = reduce(twisted[0], twisted[1]);
-            });
-          }
+      worker.on("exit", (code, signal) => {
+        if (code !== 0) {
+          console.log(`worker exited with error code: ${code}`);
+        }
+      });
 
-          callback(final);
+      cluster.on("exit", (worker, code) => {
+        if (reduce_count === reducer_task_size) {
+          // kill remaining workers
+          for (const id in cluster.workers) {
+            const process_id = cluster.workers[id].process.pid;
+            // process.kill(process_id)
+            cluster.workers[id].kill();
+          }
+          process.exit(0);
         }
       });
     }
-  } else {
-    let task = tasks[cluster.worker.id - 1];
-    let intermediate = map(task);
-    process.send({
-      from: cluster.worker.id,
-      signal: "MapCompleted",
-      intermediate,
-    });
+  } else if (cluster.isWorker) {
+    process.on("message", (msg) => {
+      if (msg["map_task"]) {
+        const mapped = map(msg["map_task"]);
+        process.send({
+          from: cluster.worker.id,
+          signal: "MapCompleted",
+          mapped,
+        });
+      }
+      if (msg["reduce_task"]) {
+        let reduced = {};
+        msg["reduce_task"].map((el) => {
+          reduced[el[0]] = reduce(el[0], el[1]);
+        });
 
-    cluster.worker.kill();
+        process.send({
+          from: cluster.worker.id,
+          signal: "ReduceCompleted",
+          reduced,
+        });
+        cluster.worker.kill();
+      }
+    });
   }
 };
 
 module.exports = function (c) {
   cores = c || default_cores;
-  return mapreduce_cluster;
+  return task_mgr;
 };
